[kernels/cpu] inherit data structure and neighbor list from singlecpu
