+++ cmake/sources/kernels/cpu.cmake
@@ -4,7 +4,6 @@ SET(CPU_INCLUDE_DIR "${READDY_GLOBAL_DIR}/kernels/cpu/include")
 # --- main sources ---
 LIST(APPEND CPU_SOURCES "${SOURCES_DIR}/CPUKernel.cpp")
 LIST(APPEND CPU_SOURCES "${SOURCES_DIR}/CPUStateModel.cpp")
-LIST(APPEND CPU_SOURCES "${SOURCES_DIR}/model/ParticleData.cpp")
 LIST(APPEND CPU_SOURCES "${SOURCES_DIR}/potentials/CPUPotentials.cpp")
 LIST(APPEND CPU_SOURCES "${SOURCES_DIR}/potentials/CPUPotentialFactory.cpp")
