+++ include/readdy/kernel/singlecpu/model/SingleCPUNeighborList.h
@@ -19,8 +19,6 @@ namespace readdy {
     namespace kernel {
         namespace singlecpu {
             namespace model {
-                
-
                 struct ParticleIndexPair {
                     size_t idx1, idx2;
                     ParticleIndexPair(size_t idx1, size_t idx2){
@@ -58,12 +56,8 @@ namespace readdy {
                     }
                 };
 
-                struct SingleCPUNeighborListBase {
-                    virtual void create(const SingleCPUParticleData &data) = 0;
-                };
-
                 template<typename container=std::unordered_set<ParticleIndexPair, ParticleIndexPairHasher>>
-                struct SingleCPUNeighborListContainer : public SingleCPUNeighborListBase {
+                struct SingleCPUNeighborListContainer {
                     typedef typename container::iterator iter_type;
                     typedef typename container::const_iterator const_iter_type;
 
@@ -74,6 +68,7 @@ namespace readdy {
                     virtual const_iter_type end() const { return cend(); };
                     virtual const_iter_type cend() const { return pairs->cend(); };
 
+                    virtual void create(const SingleCPUParticleData &data) = 0;
                 protected:
                     std::unique_ptr<container> pairs = std::make_unique<container>();
                 };
@@ -96,16 +91,111 @@ namespace readdy {
 
                 };
 
-                template<typename T=std::unordered_set<ParticleIndexPair, ParticleIndexPairHasher>>
-                struct NotThatNaiveSingleCPUNeighborList : public SingleCPUNeighborListContainer<T> {
-                    typedef SingleCPUNeighborListContainer<T> super;
+                template<typename container=std::unordered_set<ParticleIndexPair, ParticleIndexPairHasher>>
+                class NotThatNaiveSingleCPUNeighborList : public SingleCPUNeighborListContainer<container> {
+                    using super = readdy::kernel::singlecpu::model::SingleCPUNeighborListContainer<container>;
+                public:
+                    NotThatNaiveSingleCPUNeighborList(const readdy::model::KernelContext *const context) : ctx(context) {
+                    }
 
-                    NotThatNaiveSingleCPUNeighborList(const readdy::model::KernelContext *const ctx);
-                    virtual ~NotThatNaiveSingleCPUNeighborList();
+                    virtual void create(const SingleCPUParticleData &data) override {
+                        setupBoxes();
+                        fillBoxes(data);
+                    }
 
-                    virtual void create(const SingleCPUParticleData &data) override;
-                    virtual void setupBoxes();
-                    virtual void fillBoxes(const SingleCPUParticleData &data);
+                    virtual void setupBoxes() {
+                        const auto simBoxSize = ctx->getBoxSize();
+                        if (boxes.empty()) {
+                            double maxCutoff = 0;
+                            for (auto &&e : ctx->getAllOrder2RegisteredPotentialTypes()) {
+                                for (auto &&p : ctx->getOrder2Potentials(std::get<0>(e), std::get<1>(e))) {
+                                    maxCutoff = maxCutoff < p->getCutoffRadius() ? p->getCutoffRadius() : maxCutoff;
+                                }
+                            }
+                            for (auto &&e : ctx->getAllOrder2Reactions()) {
+                                maxCutoff = maxCutoff < e->getEductDistance() ? e->getEductDistance() : maxCutoff;
+                            }
+                            NotThatNaiveSingleCPUNeighborList::maxCutoff = maxCutoff;
+                            if (maxCutoff > 0) {
+
+                                for (unsigned int i = 0; i < 3; ++i) {
+                                    nBoxes[i] = (int) floor(simBoxSize[i] / maxCutoff);
+                                    if (nBoxes[i] == 0) nBoxes[i] = 1;
+                                    boxSize[i] = simBoxSize[i] / nBoxes[i];
+                                }
+                                for (long i = 0; i < nBoxes[0]; ++i) {
+                                    for (long j = 0; j < nBoxes[1]; ++j) {
+                                        for (long k = 0; k < nBoxes[2]; ++k) {
+                                            boxes.push_back({i, j, k, k + j * nBoxes[2] +
+                                                                      i * nBoxes[2] * nBoxes[1]});
+                                        }
+                                    }
+                                }
+                                for (long i = 0; i < nBoxes[0]; ++i) {
+                                    for (long j = 0; j < nBoxes[1]; ++j) {
+                                        for (long k = 0; k < nBoxes[2]; ++k) {
+                                            auto me = getBox(i, j, k);
+                                            me->addNeighbor(getBox(i + 0, j + 0, k + 1));
+                                            me->addNeighbor(getBox(i + 0, j + 1, k - 1));
+                                            me->addNeighbor(getBox(i + 0, j + 1, k + 0));
+                                            me->addNeighbor(getBox(i + 0, j + 1, k + 1));
+                                            me->addNeighbor(getBox(i + 1, j - 1, k - 1));
+                                            me->addNeighbor(getBox(i + 1, j - 1, k + 0));
+                                            me->addNeighbor(getBox(i + 1, j - 1, k + 1));
+                                            me->addNeighbor(getBox(i + 1, j + 0, k - 1));
+                                            me->addNeighbor(getBox(i + 1, j + 0, k + 0));
+                                            me->addNeighbor(getBox(i + 1, j + 0, k + 1));
+                                            me->addNeighbor(getBox(i + 1, j + 1, k - 1));
+                                            me->addNeighbor(getBox(i + 1, j + 1, k + 0));
+                                            me->addNeighbor(getBox(i + 1, j + 1, k + 1));
+                                        }
+                                    }
+                                }
+                            }
+                        }
+                    }
+                    virtual void fillBoxes(const SingleCPUParticleData &data) {
+                        const auto simBoxSize = ctx->getBoxSize();
+                        if (maxCutoff > 0) {
+
+                            for (auto &&box : boxes) {
+                                box.particleIndices.clear();
+                            }
+
+                            auto it_pos = data.cbegin_positions();
+                            unsigned long idx = 0;
+                            const auto shift = readdy::model::Vec3(.5 * simBoxSize[0], .5 * simBoxSize[1],
+                                                                   .5 * simBoxSize[2]);
+                            while (it_pos != data.cend_positions()) {
+                                const auto pos_shifted = *it_pos + shift;
+                                const long i = (const long) floor(pos_shifted[0] / boxSize[0]);
+                                const long j = (const long) floor(pos_shifted[1] / boxSize[1]);
+                                const long k = (const long) floor(pos_shifted[2] / boxSize[2]);
+                                auto box = getBox(i, j, k);
+                                if (box) {
+                                    box->particleIndices.push_back(idx);
+                                }
+                                ++idx;
+                                ++it_pos;
+                            }
+
+                            for (auto &&box : boxes) {
+                                for (long i = 0; i < box.particleIndices.size(); ++i) {
+                                    const auto pI = box.particleIndices[i];
+                                    for (long j = i + 1; j < box.particleIndices.size(); ++j) {
+                                        super::pairs->insert((*super::pairs).end(), {pI, box.particleIndices[j]});
+                                    }
+
+                                    for (auto &&neighboringBox : box.neighboringBoxes) {
+                                        for (const auto &pJ : neighboringBox->particleIndices) {
+                                            super::pairs->insert((*super::pairs).end(), {pI, pJ});
+                                        }
+                                    }
+                                }
+                            }
+                        }
+                    }
+
 
                 protected:
                     struct Box {
@@ -114,18 +204,31 @@ namespace readdy {
                         long i, j, k;
                         long id = 0;
 
-                        Box(long i, long j, long k, long id);
-                        void addNeighbor(Box *box);
+                        Box(long i, long j, long k, long id) : i(i), j(j), k(k), id(id) {
+                        }
+                        void addNeighbor(Box *box) {
+                            if (box && box->id != id) neighboringBoxes.push_back(box);
+                        }
                     };
-                    const readdy::model::KernelContext *ctx;
                     std::vector<Box> boxes{};
                     std::array<int, 3> nBoxes{{0, 0, 0}};
                     readdy::model::Vec3 boxSize{0, 0, 0};
                     double maxCutoff = 0;
 
-                    long positive_modulo(long i, long n) const;
-                    Box *getBox(long i, long j, long k);
-
+                    long positive_modulo(long i, long n) const {
+                        return (i % n + n) % n;
+                    }
+                    Box *getBox(long i, long j, long k) {
+                        const auto &periodic = ctx->getPeriodicBoundary();
+                        if (periodic[0]) i = positive_modulo(i, nBoxes[0]);
+                        else if (i < 0 || i >= nBoxes[0]) return nullptr;
+                        if (periodic[1]) j = positive_modulo(j, nBoxes[1]);
+                        else if (j < 0 || j >= nBoxes[1]) return nullptr;
+                        if (periodic[2]) k = positive_modulo(k, nBoxes[2]);
+                        else if (k < 0 || k >= nBoxes[2]) return nullptr;
+                        return &boxes[k + j * nBoxes[2] + i * nBoxes[2] * nBoxes[1]];
+                    }
+                    const readdy::model::KernelContext *const ctx;
                 };
 
                 struct SingleCPUNeighborList : public NotThatNaiveSingleCPUNeighborList<> {
